🐛 FIX: Correction architecture
