+++ utils/afficher_une.py
@@ -1,0 +1,2 @@
+def affiche_une_tache():
+    return 1
