+++ utils/ajout_tache.py
@@ -1,0 +1,2 @@
+def ajout_tache():
+    return 1
