+++ utils/supprimer_tout.py
@@ -1,0 +1,2 @@
+def supprimer_tout_tache():
+    return 1
