+++ utils/supprimer_une.py
@@ -1,0 +1,2 @@
+def supprime_tache():
+    return 1
